+++ packages/backend/jestSetup.js
@@ -1,11 +1,14 @@
-import { setEngine, CryptoEngine } from'pkijs'
+import { setEngine, CryptoEngine } from 'pkijs'
 import { Crypto } from '@peculiar/webcrypto'
 
-const crypto = new Crypto();
-global.crypto = crypto;
+const crypto = new Crypto()
+global.crypto = crypto
 
-setEngine('newEngine', new CryptoEngine({
+setEngine(
+  'newEngine',
+  new CryptoEngine({
     name: 'newEngine',
     // @ts-ignore
     crypto: crypto,
-  }))
+  })
+)