+++ packages/state-manager/src/sagas/communities/communities.slice.ts
@@ -40,6 +40,7 @@ export const communitiesSlice = createSlice({
     sendCommunityCaData: state => state,
     sendCommunityMetadata: state => state,
     createNetwork: (state, _action: PayloadAction<CreateNetworkPayload>) => state,
+    joinNetwork: (state, _action: PayloadAction<InvitationData>) => state,
     storePeerList: (state, action: PayloadAction<StorePeerListPayload>) => {
       communitiesAdapter.updateOne(state.communities, {
         id: action.payload.communityId,
@@ -50,7 +51,7 @@ export const communitiesSlice = createSlice({
     },
     resetApp: (state, _action) => state,
     launchCommunity: (state, _action: PayloadAction<string>) => state,
-    customProtocol: (state, _action: PayloadAction<InvitationData>) => state,
+    customProtocol: (state, _action: PayloadAction<string[]>) => state,
     setInvitationCodes: (state, action: PayloadAction<InvitationPair[]>) => {
       state.invitationCodes = action.payload
     },