+++ packages/types/src/socket.ts
@@ -14,6 +14,7 @@ export enum SocketActionTypes {
   COMMUNITY_LAUNCHED = 'communityLaunched',
   COMMUNITY_METADATA_STORED = 'communityMetadataStored',
   CREATE_COMMUNITY = 'createCommunity',
+  DOWNLOAD_INVITE_DATA = 'downloadInviteData',
   LAUNCH_COMMUNITY = 'launchCommunity',
   LEAVE_COMMUNITY = 'leaveCommunity',
   SET_COMMUNITY_CA_DATA = 'setCommunityCaData',