+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 [unreleased]
 
+# Features:
+
+* Add support for new format of invitation link: `c=<cid>&t=<token>&s=<serverAddress>&i=<inviterAddress>`
+
 # Refactorings:
 
 * Use ack for CREATE_NETWORK and simplify
@@ -10,6 +14,10 @@
 * Fixes channel name creation logic
 * Allow JPEG and GIF files as profile photos ([#2332](https://github.com/TryQuiet/quiet/issues/2332))
 
+# Other:
+
+* Refactored package.json to have consistent license "GPL-3.0-or-later"
+ 
 [2.1.2]
 
 # Refactorings: